Thumbnail images in gallery
sorl-thumbnail uses os.path.join() on the MEDIA_URL and the path of the
photo. This caused issues when the photo path came from elasticsearch
because it was a relative path that started with "/"
("/media/"), which caused it to treat is as a full path
then, Django pooped out because it looked like
an aboslute path to somewhere on the filesystem.
My solution is to remove the /media prefix if
it comes from elasticsearch, but this may not work
when things are moved to AWS